Use random port for Devtools integration tests
Closes gh-15013

+++ spring-boot-tests/spring-boot-integration-tests/spring-boot-devtools-tests/src/test/java/org/springframework/boot/devtools/tests/RemoteApplicationLauncher.java
@@ -20,6 +20,7 @@ import java.io.File;
 import java.io.FileReader;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.function.BiFunction;
 
 import org.springframework.boot.devtools.RemoteSpringApplication;
 import org.springframework.boot.devtools.tests.JvmLauncher.LaunchedJvm;
@@ -40,26 +41,46 @@ abstract class RemoteApplicationLauncher implements ApplicationLauncher {
 			throws Exception {
 		LaunchedJvm applicationJvm = javaLauncher.launch("app",
 				createApplicationClassPath(), "com.example.DevToolsTestApplication",
-				"--server.port=12345", "--spring.devtools.remote.secret=secret");
-		awaitServerPort(applicationJvm.getStandardOut());
-		LaunchedJvm remoteSpringApplicationJvm = javaLauncher.launch(
-				"remote-spring-application", createRemoteSpringApplicationClassPath(),
-				RemoteSpringApplication.class.getName(),
-				"--spring.devtools.remote.secret=secret", "http://localhost:12345");
-		awaitRemoteSpringApplication(remoteSpringApplicationJvm.getStandardOut());
+				"--server.port=0", "--spring.devtools.remote.secret=secret");
+		int port = awaitServerPort(applicationJvm.getStandardOut());
+		BiFunction<Integer, File, Process> remoteRestarter = getRemoteRestarter(
+				javaLauncher);
 		return new LaunchedApplication(new File("target/remote"),
 				applicationJvm.getStandardOut(), applicationJvm.getStandardError(),
-				applicationJvm.getProcess(), remoteSpringApplicationJvm.getProcess());
+				applicationJvm.getProcess(), remoteRestarter.apply(port, null),
+				remoteRestarter);
+	}
+
+	private BiFunction<Integer, File, Process> getRemoteRestarter(
+			JvmLauncher javaLauncher) {
+		return (port, classesDirectory) -> {
+			try {
+				LaunchedJvm remoteSpringApplicationJvm = javaLauncher.launch(
+						"remote-spring-application",
+						createRemoteSpringApplicationClassPath(classesDirectory),
+						RemoteSpringApplication.class.getName(),
+						"--spring.devtools.remote.secret=secret",
+						"http://localhost:" + port);
+				awaitRemoteSpringApplication(remoteSpringApplicationJvm.getStandardOut());
+				return remoteSpringApplicationJvm.getProcess();
+			}
+			catch (Exception ex) {
+				throw new IllegalStateException(ex);
+			}
+		};
 	}
 
 	protected abstract String createApplicationClassPath() throws Exception;
 
-	private String createRemoteSpringApplicationClassPath() throws Exception {
-		File remoteDirectory = new File("target/remote");
-		FileSystemUtils.deleteRecursively(remoteDirectory);
-		remoteDirectory.mkdirs();
-		FileSystemUtils.copyRecursively(new File("target/test-classes/com"),
-				new File("target/remote/com"));
+	private String createRemoteSpringApplicationClassPath(File classesDirectory)
+			throws Exception {
+		if (classesDirectory == null) {
+			File remoteDirectory = new File("target/remote");
+			FileSystemUtils.deleteRecursively(remoteDirectory);
+			remoteDirectory.mkdirs();
+			FileSystemUtils.copyRecursively(new File("target/test-classes/com"),
+					new File("target/remote/com"));
+		}
 		List<String> entries = new ArrayList<>();
 		entries.add("target/remote");
 		for (File jar : new File("target/dependencies").listFiles()) {
